+++ package.js
@@ -11,4 +11,16 @@ Package.on_use(function (api) {
   api.add_files('timesync-client.js', 'client');
 
   api.export('TimeSync', 'client');
+  api.export('SyncInternals', 'client', {testOnly: true} );
+});
+
+Package.on_test(function (api) {
+  api.use([
+    'tinytest',
+    'test-helpers'
+  ]);
+
+  api.use('timesync');
+
+  api.add_files('tests/client.js', 'client');
 });